feat: add userAvatar model
model 'userAvatar' added, which is a junction table to allow M:N - a user can have many avatars, and an avatar can belong to many users. added now directly; before, it was being created automatically by sequelize based on the string passed to 'through' in the user/avatar relationships options.